Separated bugzilla functionality into separate classes
Created a BugzillaCommenter class for commenting on issues after
changes are merged.
Created a BugzillaPrChecker class for changing the bugzilla ticket
status and adding PR comments to the ticket.
Created a BugzillaCommon module which contains shared functionality.
